items: avoid interpolating inactive items
This covers not interpolating items that are inactive, such as stopped
boulders. An item may still be active in the item list such that it
retains collision, but inanimate, sich as boulders or mines, so we need
to check both statuses. The TR1 test here has also been simplified to
match.